Implemented missing options
Added most of the old options in the config file and implemented their features

+++ src/main/java/com/gilles_m/rpg_regen/CombatChecker.java
@@ -18,7 +18,7 @@ public class CombatChecker {
         this.player = player;
     }
 
-    public void start(final float delay) {
+    public void start(final double delay) {
         if(task != null) {
             task.cancel();
         }
